Make this compatible with the MassMessage multilingual refactor
This makes this extension compatible with the MessageBuilder
changes in 0929da16d59d0f43.

For this to work, this changes the required patch to MassMessageJob
to pass more things in the hook. I.E:
if ( !\Hooks::run( 'MassMessageJobBeforeMessageSent',
 [ $this, $targetPage, $subject, $message, $pageSubject, $pageMessage, $comment ] ) ) {

Additionally, this sets load_composer_autoload so that the composer
class for converting html to text is properly loaded.

This also simplifies that code related to generating absolute urls
to use the new absoluteURLs parser output option that was introduced
in 1.38 (And hence bumping the MediaWiki requirement to 1.38).

Change-Id: I18544ac344d023722ea67926c73abde241af2233
(cherry picked from commit a3eed9a3cce7477bbfee53e56086e77cf9c483ef)